figured out how to setState on a nested Object in react.
Able to change a single task.editState on click. Now that we can change the state, it then renders an input area instead of task.text on the screen.

Now to add the user to submit the new text, from the input, and change the 'edit' to be 'submit'. Basically we will need to make a mini form I think for the task.editState === true;